LayoutTests/platform/mac/editing/spelling/autocorrection-blockquote-crash-expected.txt

This test checks that markers are correct when auto correcting in the blockquote. If you type "l" and " ", there should be blue dots under the word: "notational". This test also makes sure that the replacement does not cause breaking out of blockquotes.

PASS internals.hasAutocorrectedMarker(document, 5, 10) is true
PASS successfullyParsed is true

TEST COMPLETE

would this 
test notational 
make a difference?
